Reorganize main form sources
