+++ Timelapse/Application/MainForm/ControlDefinitions.h
