+++ Timelapse/Application/MainForm/Designer.cpp
@@ -1,4 +1,4 @@
-#include "MainForm.Internal.h"
+#include "SharedState.h"
 
 using namespace Timelapse;
