+++ Timelapse/Application/MainForm/EventHandlers.h
