+++ Timelapse/Application/MainForm/MainForm.resx
