+++ Timelapse/Application/MainForm/TestingUtilities.cpp
@@ -1,4 +1,4 @@
-#include "MainForm.Internal.h"
+#include "SharedState.h"
 
 using namespace Timelapse;
 using namespace Timelapse::Logging;
