+++ Timelapse/Application/Settings/Settings.h
