+++ Timelapse/Timelapse.vcxproj.filters
@@ -8,6 +8,12 @@
     <Filter Include="Source Files\Application">
       <UniqueIdentifier>{20F7E82E-20E4-4F01-89C5-4E3C82AF31C6}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Source Files\Application\MainForm">
+      <UniqueIdentifier>{D5B2E6B4-7B69-4F0C-9D84-4439B3DE4394}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\Application\Settings">
+      <UniqueIdentifier>{A9E5C836-661F-4794-8A2C-66D0805128A0}</UniqueIdentifier>
+    </Filter>
     <Filter Include="Source Files\Core">
       <UniqueIdentifier>{CE7B21B7-18BC-4E3B-8B7E-95E6AEB7C9B8}</UniqueIdentifier>
     </Filter>
@@ -23,6 +29,15 @@
     </Filter>
     <Filter Include="Header Files\Application">
       <UniqueIdentifier>{B6F71A0B-4F23-4CF7-8E5A-B019A765D17A}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\Application\MainForm">
+      <UniqueIdentifier>{AB96C3A6-6D7F-4D10-A5E5-D577BC08A4A5}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\Application\Settings">
+      <UniqueIdentifier>{0AD3CE07-0C3B-4E39-A43A-8E430F6F4E4D}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\Application\Macros">
+      <UniqueIdentifier>{43A1211E-5D5B-4C83-9F4B-2B7970D6D24F}</UniqueIdentifier>
     </Filter>
     <Filter Include="Header Files\Core">
       <UniqueIdentifier>{96D579D9-49AB-4F3A-A1FA-3F7E6DAEB8A2}</UniqueIdentifier>
@@ -54,26 +69,26 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Application\MainForm.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\MainForm\MainForm.h">
+      <Filter>Header Files\Application\MainForm</Filter>
     </ClInclude>
-    <ClInclude Include="Application\MainForm.Controls.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\MainForm\ControlDefinitions.h">
+      <Filter>Header Files\Application\MainForm</Filter>
     </ClInclude>
-    <ClInclude Include="Application\MainForm.EventHandlers.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\MainForm\EventHandlers.h">
+      <Filter>Header Files\Application\MainForm</Filter>
     </ClInclude>
-    <ClInclude Include="Application\MainForm.Internal.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\MainForm\SharedState.h">
+      <Filter>Header Files\Application\MainForm</Filter>
     </ClInclude>
-    <ClInclude Include="Application\MainForm.MapRusher.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\MainForm\MapRusherFeatures.h">
+      <Filter>Header Files\Application\MainForm</Filter>
     </ClInclude>
-    <ClInclude Include="Application\Macro.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\Macros\Macro.h">
+      <Filter>Header Files\Application\Macros</Filter>
     </ClInclude>
-    <ClInclude Include="Application\Settings.h">
-      <Filter>Header Files\Application</Filter>
+    <ClInclude Include="Application\Settings\Settings.h">
+      <Filter>Header Files\Application\Settings</Filter>
     </ClInclude>
     <ClInclude Include="Core\Inventory.h">
       <Filter>Header Files\Core</Filter>
@@ -122,35 +137,35 @@
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Application\MainForm.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\MainForm.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Bots.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\BotsTab.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Designer.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\Designer.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Filters.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\FiltersTab.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Hacks.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\HacksTab.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.MapRusher.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\MapRusherFeatures.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Packets.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\PacketTools.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Testing.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\TestingUtilities.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\MainForm.Validation.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\MainForm\ValidationHelpers.cpp">
+      <Filter>Source Files\Application\MainForm</Filter>
     </ClCompile>
-    <ClCompile Include="Application\Settings.cpp">
-      <Filter>Source Files\Application</Filter>
+    <ClCompile Include="Application\Settings\Settings.cpp">
+      <Filter>Source Files\Application\Settings</Filter>
     </ClCompile>
     <ClCompile Include="Core\Inventory.cpp">
       <Filter>Source Files\Core</Filter>
@@ -166,7 +181,7 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Application\MainForm.resx">
+    <EmbeddedResource Include="Application\MainForm\MainForm.resx">
       <Filter>Resource Files\Application</Filter>
     </EmbeddedResource>
   </ItemGroup>
